+++ src/geo/projection/vertical_perspective_transform.ts
@@ -304,6 +304,7 @@ export class VerticalPerspectiveTransform implements ITransform {
             clippingPlane: this._cachedClippingPlane as [number, number, number, number],
             projectionTransition: applyGlobeMatrix ? 1 : 0,
             fallbackMatrix: this._globeViewProjMatrix32f,
+            clipAntimeridian: overscaledTileID?.canonical.z === 0,
         };
     }